Feat: 고민글 게시판 필터링, 서버 연결, 컴포넌트 분리
